ddl: split testIntegrationSuite7 to another file
